add 3dconv model filelists supports returning sequences of images added super model that carries information used during training to infer correctness
everything seems to work so far
