+++ src/forgery_detection/lightning/benchmark.py
@@ -4,10 +4,10 @@ from pathlib import Path
 import click
 import torch
 
+from forgery_detection.lightning.logging.utils import SystemMode
 from forgery_detection.lightning.system import Supervised
 from forgery_detection.lightning.utils import get_latest_checkpoint
 from forgery_detection.lightning.utils import PythonLiteralOptionGPUs
-from forgery_detection.lightning.utils import SystemMode
 
 
 @click.command()
